fix(ci): python-full-coverage — split oversized test files and fix salvage stub
Two test files exceeded the 1500-line maintainability limit; split them into
part_037/part_019. The salvage regression stub was missing _worktrees_root,
so the test hit AttributeError before exercising the mocked capture path.
